+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - Theme: Supports passing multiple theme fragments to the same provider
 - FormControl: Label margin-bottom: 0 for the last child for better composition
 - DropdownMenu: Fixed items not working inside router links
+- Tooltip: Only apply aria-describedby when tooltip is active